requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
torch==2.1.1
transformers==4.35.2
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6 